Fix truncation and partitioning in number_to_currency helper
Instead of calling `divmod` on the float and running the risk of a
floating point error, cast to string and split on the decimal.

Once split into a string, all formatting is done on the string. To
ensure that the string has enough digits for precision, automatically
add enough zeros at the end.